Extract modules and cleanup
