+++ src/ast/mod.rs
@@ -1,3 +1,5 @@
 pub mod expr;
+pub mod literal;
+pub mod name;
 pub mod stmt;
 pub mod token;
