+++ src/interpreter/literal.rs
@@ -1,5 +1,6 @@
+use crate::ast::literal::Literal;
+
 use super::{value::Value, Result};
-use crate::ast::expr::Literal;
 pub(super) trait LiteralInterpreter {
     fn interpret(&self) -> Result<Value>;
 }
